requirements.txt
Flask-Login~=0.6.3
Flask-Migrate~=4.1.0
SQLAlchemy~=2.0.40
requests~=2.32.3
pandas~=2.2.3
Flask~=2.3.3
Werkzeug~=3.1.3
#for later
psycopg2-binary~=2.9.10
psycopg[binary]~=3.2
python-dotenv~=0.21.1
reportlab~=4.4.1
alembic~=1.15.2
PyJWT~=2.10.1
flask-cors~=5.0.1
openpyxl~=3.1.5
gunicorn==21.2.0
flasgger~=0.9.7.1
boto3~=1.40.55
pillow~=11.2.1
botocore~=1.40.55
python-dateutil~=2.9.0.post0
APScheduler~=3.11.0
flask-socketio
gevent
gevent-websocket
redis~=7.1.0